aya-integrations/backend/requirements.txt
Flask==3.1.0
flask-cors
flask-sock
torch
torchaudio
numpy
librosa
webrtcvad
torch
ffmpeg-python
transformers==4.40.0
huggingface_hub==0.21.3
tokenizers
datasets>=2.18.0
git+https://github.com/facebookresearch/seamless_communication.git
fairseq2
simuleval
setuptools 
wheel
accelerate
faster-whisper
python-dotenv